requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
openai>=1.0.0
arxiv==2.1.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
sse-starlette==1.8.2
tenacity==8.2.3
python-dotenv==1.0.0 